+++ frontend/app/components/templates/FacePaymentPage/FacePaymentPage.tsx
@@ -16,6 +16,8 @@ import styles from './FacePaymentPage.module.scss'
 import Feedback from '@/modules/Feedback/Feedback'
 import { ActivationForm } from '@/modules/ActivationForm/ActivationForm'
 import { PaymentPay } from '@/modules/PaymentPay/PaymentPay'
+import { PaymentMenu } from '@/modules/PaymentMenu/PaymentMenu'
+import { checkAccount } from '@/api/paymentAPI'
 
 type Props = {
   id: number
@@ -23,6 +25,11 @@ type Props = {
   short_description: string
   long_description: string
 }
+
+const mainMenuInfo = [
+  'Before we start our payment, create your account first',
+  'Do your transactions using face payment'
+]
 
 export const FacePaymentPage = ({
   id,
@@ -35,6 +42,7 @@ export const FacePaymentPage = ({
   const [openModal, setOpenModal] = useState(false)
   const [cart, setCart] = useState<ShoppingItem>()
   const [total, setTotal] = useState(0)
+  const [isAccountMade, setIsAccountMade] = useState(false)
   const { session_id } = parseCookies()
 
   const moveStep = (numStep: number, moveStepStepper?: boolean) => {
@@ -64,6 +72,40 @@ export const FacePaymentPage = ({
       }
     }
   }
+
+  const checkStatus = async () => {
+    if (session_id) {
+      try {
+        const res = await checkAccount(session_id)
+        if (res) {
+          // Current userflow, not activated is considered same as not registered
+          console.log(res)
+          setIsAccountMade(res?.have_active_account)
+        }
+      } catch (e) {
+        console.error(e)
+      }
+    } else {
+      setOpenModal(true)
+    }
+  }
+
+  const menuButtons = [
+    {
+      title: 'Account Registration',
+      onClick: () => {
+        setCurrentStep(3)
+        setCurrentStepStepper(2)
+      }
+    },
+    {
+      title: 'Face Payment',
+      onClick: () => {
+        setCurrentStep(5)
+        setCurrentStepStepper(4)
+      }
+    }
+  ]
 
   return (
     <>
@@ -97,13 +139,26 @@ export const FacePaymentPage = ({
               Please access this demo via smartphone or any device with at least
               HD camera resolution for better performance and experience
             </p>
-            <Button color={Color.Primary} onClick={() => moveStep(1, true)}>
+            <Button
+              color={Color.Primary}
+              onClick={() => {
+                moveStep(1, false)
+                checkStatus()
+              }}>
               Start
             </Button>
           </div>
         )}
 
         {currentStep === 2 && (
+          <PaymentMenu
+            buttons={menuButtons}
+            disabledList={[+!isAccountMade]}
+            title={mainMenuInfo[+isAccountMade]}
+          />
+        )}
+
+        {currentStep === 3 && (
           <FaceRegistration
             onChecking={() => {
               createVisitorActivities(id, session_id, 1)
@@ -116,16 +171,17 @@ export const FacePaymentPage = ({
           />
         )}
 
-        {currentStep === 3 && (
+        {currentStep === 4 && (
           <ActivationForm
             nextStep={() => {
               createVisitorActivities(id, session_id, 3)
-              moveStep(1, true)
+              checkStatus()
+              setCurrentStep(2)
             }}
           />
         )}
 
-        {currentStep === 4 && (
+        {currentStep === 5 && (
           <Catalog
             onAddToCart={(item) => {
               setCart(item)
@@ -135,7 +191,7 @@ export const FacePaymentPage = ({
           />
         )}
 
-        {currentStep === 5 && cart && (
+        {currentStep === 6 && cart && (
           <Cart
             onBack={() => moveStep(-1)}
             onCheckout={(item) => {
@@ -146,7 +202,8 @@ export const FacePaymentPage = ({
             item={cart}
           />
         )}
-        {currentStep === 6 && cart && (
+
+        {currentStep === 7 && cart && (
           <OrderSummary
             cart={cart}
             onNext={(total) => {
@@ -160,6 +217,7 @@ export const FacePaymentPage = ({
         {currentStep === 7 && (
           <PaymentPay sessionId={session_id} amount={total} />
         )}
+
         {currentStep === 8 && (
           <Feedback
             id={id}