+++ frontend/app/types/responses.ts
@@ -162,3 +162,8 @@ export type CheckLimitResponse = {
     }
   ]
 }
+export type CheckAccountResponse = {
+  have_active_account: boolean,
+  message: string,
+  ok: boolean
+}